requirements.txt
Babel==2.7.0
dataclasses==0.6
Django==2.2.4
django-jalali==3.1.0
django-phonenumber-field==3.0.1
django-redis==4.10.0
django-rest-framework==0.1.0
djangorestframework==3.10.2
djongo==1.2.33
jdatetime==3.5.0
phonenumbers==8.10.16
Pillow==6.1.0
pymongo==3.9.0
pytz==2019.2
redis==3.3.7
sqlparse==0.2.4
